Fix address conversion in qspi driver
Existing address format in ltq_spinand & qspi driver is
inconsistent as it somehow uses different endianness, depending
on quad or single mode. Eventually this causes qspi driver
issue if we use single mode.

This change is to make the endianness consistent, hence removing
dependency on quad or single mode. Qspi driver is also modified
to fix bugs that was hidden before and uncovered now.