PONRTSYS-1977: gpio: intel-sso: Deactivate not used SOUT groups
This will make the driver actively deactivate the SOUT groups which are not
used and not only activate the needed groups.

The reset value of this register is 0, but the U-Boot can also use the SSO
registers and could activate all groups.
This is needed to restrict the SSO to 2 groups on the PRX321 SFU board.