DRVLIB_SW-846 - Add P31G into lantiq phy driver list
This is needed so that P31G can be hooked into linux PHY framework.
There are no known differences in register mapping with P11G,
so for now I simply extend existing lantiq.c (which is used in
grx500 platform). This is to avoid the needs of rewriting a new
driver/code.