PONRTSYS-3635: wait only 0.1ms to 0.2ms for PPv4 FW answer
If the PPv4 FW didn't answer immediately to a request the PPv4 driver used
msleep(10) to wait for the next try. msleep() with values less than 20 is
not working well and mostly the CPU will wait around 20ms.
Use usleep_range() now and check for the answer ever 0.1ms to 0.2ms, this
is able to wait for such a short time. This way we get the answer from the
PPv4 FW much earlier and we can finish the operations much faster. This
gives a significant performance improvement to the OMCI configuration.

The number of tries was also increased so the max wait time should be similar to before.