net: sched: add simple drr offload support
The commit adds simple offload support to the tc-drr. The offload can
be used to support scheduler and its attached queues configuration.

Example configuration commands are shown beloW:

  tc qdisc add dev eth0_0 root handle 8001: drr
  tc class add dev eth0_0 parent 8001: classid 8001:1 drr quantum 4096
  tc class add dev eth0_0 parent 8001: classid 8001:2 drr quantum 8192
  tc class add dev eth0_0 parent 8001: classid 8000:3 drr quantum 8192

The first command passes the parent id and the handle to the
offload hook. These are the root(0xffffffff) and the qdisc handle id.
The driver offloading the drr has to allocate and configure a scheduler
resource in this case. The same is valid even if the drr is not the root
qdisc like below:

Use case 1:                   Use case 2:

  root = drr                      root
   / |  \                            \
  q1 q2 q3                           drr
                                    / | \
                                   q1 q2 q3

The commands creating the classes are passing the necessary parameters
to offload hook for the queue resource allocations and configuration.
The parent id in this case for each class is equal to 8001. In this way
the driver can recognize that it has to configure a queue.

Signed-off-by: Zahari Doychev <zahari.doychev@intel.com>