DRVLIB_SW-2236 - Dynamically set gphy firmware name according to SoC version
We do not have different device tree for A1/B0, yet we need
to support both under the same image. This becomes problematic
for gphy as the firmwares for A1 and B0 are unfortunately not
interchangeable.

Due to this requirement, we implement a SoC version detection
to load the correct firmware accordingly. Firmware filenames
are now hardcoded inside driver (instead of device tree).