PONRTSYS-4307: Request PTP clock in PTP driver
The PTP driver should request the PTP clock otherwise it could be
deactivated when we do not set clk_ignore_unused anymore.

The driver also sets the frequency of the clock and this frequency can
be specified in device tree, this allows it to set it to 312.5MHz in
the low power mode.