PONRTSYS-3169: grx500_wdt: Prevent race condition in VPE_OTHER accesses
The other VPE register is also used in the GIC driver, in the next access we
do some settings for the VPE configured before. If this is intercepted by an
IRQ the watchdog will be configured for the  wrong CPU.
Deactivate IRQs while we modify this to make this atomic, this should fix
some race conditions.

This hopefully fixes the sporadic watchdog triggers we see sometimes.