PONRTSYS-4774: Add padding also on the vUNI interface
When the destination MAC was already learned by the GSWIP the packets will take
the HW path on the bridge between the vUNI(PON) port and the pmapper. We will
not add any additional padding on this HW path and the packets will be dropped
somewhere. To make sure the padding is added already add the padding on
the vUNI driver.